pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "ethical-ai-validator"
dynamic = ["version"]
description = "Comprehensive ethical AI validation and auditing package"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Ethical AI Team", email = "muhammadabdullahinbox@gmail.com"}
]
maintainers = [
    {name = "Ethical AI Team", email = "muhammadabdullahinbox@gmail.com"}
]
keywords = ["ethical-ai", "bias-detection", "fairness", "ai-auditing", "gdpr-compliance"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 2.7",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Testing",
]
requires-python = ">=2.7"
dependencies = [
    "numpy>=1.16.0",
    "pandas>=0.20.0",
    "scikit-learn>=0.20.0",
    "reportlab>=3.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=4.0.0",
    "pytest-cov>=2.0.0",
    "pytest-mock>=2.0.0",
    "black>=19.0.0",
    "flake8>=3.0.0",
    "mypy>=0.700",
    "pre-commit>=1.0.0",
]
docs = [
    "sphinx>=2.0.0",
    "sphinx-rtd-theme>=0.4.0",
]
full = [
    "matplotlib>=2.0.0",
    "seaborn>=0.9.0",
    "jupyter>=1.0.0",
    "notebook>=5.0.0",
]

[project.scripts]
ethical-ai-validator = "ethical_ai_validator.cli:main"

[project.urls]
Homepage = "https://github.com/ethical-ai-validator/ethical-ai-validator"
Documentation = "https://ethical-ai-validator.readthedocs.io/"
Repository = "https://github.com/ethical-ai-validator/ethical-ai-validator"
"Bug Tracker" = "https://github.com/ethical-ai-validator/ethical-ai-validator/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
write_to = "src/ethical_ai_validator/_version.py" 